[CHANGE] Change new event style.
