+++ .travis.yml
@@ -5,7 +5,6 @@ sudo: false
 cache:
   directories:
     - $HOME/.composer/cache
-    - /home/travis/.composer/cache
 
 php:
   - 5.3
