+++ config.yml
@@ -1,5 +1,5 @@
 name: OrderPdf
-event: OrderPdf
+event: OrderPdfEvent
 code: OrderPdf
 version: 1.0.0
 service:
